+++ src/model/utils/update.js
@@ -1,7 +1,7 @@
 import DOM from "../../constants/domSelectors";
 /* eslint max-len: 0 */
 
-export const version = `0.6`;
+export const version = `0.65`;
 
 export const welcomeText = {
   title: "WykopHelper zainstalowany!",
@@ -13,28 +13,24 @@ export const welcomeText = {
 export const updateText = {
   title: "WykopHelper zaktualizowany!",
   content: `
-Dodatek WykopHelper został właśnie zaktualizowany do wersji ${version}. Wprowadzone zmiany to: <br>
+Dodatek WykopHelper został właśnie zaktualizowany do wersji <strong>${version}</strong>. Wprowadzone zmiany to: <br>
 <ul class="${DOM.MODAL.CLASSNAME.LIST}">
   <li class="${DOM.MODAL.CLASSNAME.LIST_ITEM}">
-    Dodano możliwość zmiany tekstu na odznace każdego użytkownika z osobna. By zmienić tekst wystarczy kliknąć na odznace przy danym userze, odszukać nowe pole tekstowe i wpisać tam, co dusza zapragnie :)
+    Funkcja ostrzegania przed znaleziskami podejrzanymi o szerzenie propagandy rosyjskiej została zmodyfikowana. Od teraz możesz samodzielnie ustalić, czy takie ostrzeżenie ma w ogóle być pokazywane, a także jaka ma być jego treść i dla jakich domen ma się aktywować. Zdecydować o tym możesz oczywiście w ustawieniach (ikona zębatki przy odpowiednim checkboxie). Domyślnie dodane są tam domeny podejrzane o szerzenie rosyjskiej propagandy; źródła, które posłużyły do jej skomponowania znajdują się <a href="https://plwpl.github.io/WykopHelper/propaganda.html" target="_blank">tutaj</a>.
   </li>
   <li class="${DOM.MODAL.CLASSNAME.LIST_ITEM}">
-    Pole tekstowe wyświetlające tekst komentarza w popupie odznaki uzyskało możliwość przewijania. To oznacza, że teraz bardzo długie komentarze nie będą rozciągać okna popupu nawet poza monitor.
+    Funkcja ostrzegająca przed zamknięciem strony gdy wykryte zostanie pisanie komentarza <em>powinna</em> już działać poprawnie.
   </li>
   <li class="${DOM.MODAL.CLASSNAME.LIST_ITEM}">
-    W ustawieniach można teraz zadecydować o ukrywaniu woodle (czyli wykopowej wersji doodle - okolicznościowy obrazek umieszczany na belce menu).
+    Pojawiła się opcja wyłączenia komentarzy we <strong>wszystkich</strong> znaleziskach. Teraz możesz zdecydować, czy komentarze wyłączasz globalnie, tylko w wybranych (poprzez tagi) znaleziskach, czy nigdzie. Domyślnie opcja oczywiście nie jest włączona.
   </li>
   <li class="${DOM.MODAL.CLASSNAME.LIST_ITEM}">
-    Chcesz widzieć znaleziska z określonych kategorii (np. #polityka), ale dla własnego komfortu psychicznego preferujesz nie widzieć komentarzy pod nim? Od teraz możesz zdefiniować w ustawieniach listę tagów, dla których komentarze pod znaleziskiem będą usuwane.
+    Od teraz odznaka widoczna będzie również w profilu użytkownika, między avatarem a nickiem.
   </li>
   <li class="${DOM.MODAL.CLASSNAME.LIST_ITEM}">
-    Na stronie ustawień pojawiły się linki do historii zmian oraz do strony opisującej ficzery dodatku.
-  </li>
-  <li class="${DOM.MODAL.CLASSNAME.LIST_ITEM}">
-    Kilka pomniejszych fixów i ulepszeń.
+    Drobne poprawki stylistyczne tu i ówdzie (np. nowy kolor przycisku "zapisz" w popupie odznaki). Redesign całego popupu wkrótce, a nieco później - ujednolicenie UI tak, żeby nie wyglądało jak naprędce sklejone kawałki wszystkiego ;)
   </li>
 </ul>
-🎉🎉 <strong>Szczęśliwego Nowego Roku!</strong> 🎉🎉
 `,
   button: "Okej!",
 };